Fix Seg fault and algorithm logic
Change to split child path taking so when we take left child of one node
we take the right of the other node and vice verse thus we will
be in the same distance from the center. Check if tree is empty
before accessing root element to avoid Segmentation fault.